check for package availability before attempting to install
This patch makes it clear that it's not the EPEL repo at fault.
This is needed because otherwise, the playbook would just fail
without leaving a clue of what went wrong. So to speak, a helpful
message as to whether that package (example: libdbi-dbd-mysql, etc..)
should officially be available, or if it is available through EPEL.

Change-Id: I8b2633be12a3cf780aab213a44de8022b84ef3d3
Signed-off-by: Archit Sharma <arcsharm@redhat.com>